+++ ingenialink/exceptions.py
@@ -58,5 +58,13 @@ class ILDictionaryParseError(ILError):
     """IngeniaLink dictionary parse error."""
 
 
+class ILConfigurationFileParseError(ILError):
+    """IngeniaLink configuration file parse error."""
+
+
 class ILWrongWorkingCountError(ILError):
     """PDOs process data working count expected and received differ."""
+
+
+class ILEcatStateError(ILError):
+    """IngeniaLink Ethercat state error."""