+++ tests/ethercat/test_ethercat_dictionary_v3.py
@@ -109,14 +109,14 @@ def test_read_xdf_register():
 
 
 @pytest.mark.no_connection
-def test_child_registers():
+def test_object_registers():
     dictionary_path = join_path(path_resources, dict_ecat_v3)
     ethercat_dict = DictionaryV3(dictionary_path, Interface.ECAT)
-    reg_list = ethercat_dict.child_registers("CIA301_COMMS_RPDO1_MAP", 0)
+    canopen_object = ethercat_dict.get_object("CIA301_COMMS_RPDO1_MAP", 0)
     reg_subindex = [0, 1]
     reg_uids = ["CIA301_COMMS_RPDO1_MAP", "CIA301_COMMS_RPDO1_MAP_1"]
     reg_index = [0x1600, 0x1600]
-    for index, reg in enumerate(reg_list):
+    for index, reg in enumerate(canopen_object):
         assert isinstance(reg, CanopenRegister)
         assert reg.idx == reg_index[index]
         assert reg.identifier == reg_uids[index]
@@ -124,11 +124,11 @@ def test_child_registers():
 
 
 @pytest.mark.no_connection
-def test_child_registers_not_exist():
+def test_object_not_exist():
     dictionary_path = join_path(path_resources, dict_ecat_v3)
     ethercat_dict = DictionaryV3(dictionary_path, Interface.ECAT)
     with pytest.raises(KeyError):
-        ethercat_dict.child_registers("NOT_EXISTING_UID", 0)
+        ethercat_dict.get_object("NOT_EXISTING_UID", 0)
 
 
 @pytest.mark.no_connection