+++ tests/ethernet/test_ethernet_dictionary_v2.py
@@ -118,11 +118,11 @@ def test_read_xdf_register():
 
 
 @pytest.mark.no_connection
-def test_child_registers_not_exist():
+def test_object_not_exist():
     dictionary_path = join_path(path_resources, "test_dict_eth.xdf")
     ethernet_dict = EthernetDictionaryV2(dictionary_path)
     with pytest.raises(KeyError):
-        ethernet_dict.child_registers("NOT_EXISTING_UID", 0)
+        ethernet_dict.get_object("NOT_EXISTING_UID", 0)
 
 
 @pytest.mark.no_connection